Split package into subpackages
